// PhotoStudioWebApp/PhotoStudioWebApp/ClientApp/src/app/models/room.dto.ts
export interface RoomDto {
  id: string;
  name: string;
  description: string;
  additionalInformation: string;
  booking?: BookingDto[];
  images?: RoomImageDto[];
}

export interface RoomImageDto {
  id: string;
  imagePath: string;
  imageName: string;
  roomId: string;
}

export interface BookingDto {
  id: string;
  bookFrom: Date;
  bookTo: Date;
  roomId: string;
  userId: string;
}